Removing unused orientation code from the keil project
we cleaned up the code on the Orientation folder to remove unused and unclean implementations
of the Madgwick and Kalman sensor fusion algorithms, if need a new implementation based on
the original one should be used instead. The Keil project was unaware of this change and broke
the build, this updates the project.